requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
2captcha-python==1.2.1
python-dotenv==1.0.0
python-telegram-bot==21.0.1
